refactor(test): audit pipeline.rs — dissolve file, rehome tests
- Delete test_root_compression_from_out_of_range_delete (duplicate of
  adversarial.rs test_out_of_range_root_structure_change)
- Move 4 edge-case tests to edge_cases.rs; rename db_a/db_b to
  source/target; remove stale line-number refs from comments
- Move 2 crafted-proof tests to attack.rs with crafted prefix
- Delete pipeline.rs and remove mod pipeline from mod.rs
